ClasssBundle now can generate non-overlapping class bundles of classes.
- GenerateClassBundlesFromClasses takes a list of classes and then
groups it into class bundles in such a way that each bundle has at most
1 lecture, 1 discussion, 1 problem and 1 lab class. Besides, it is
guaranteed that there will not be any overlapping classes within each
class bundle generated using this method.

- isCompatibleWith and areAllCompatible functions are added to Class
so as to check whether one class is compatible with another one and for
checking whether there is any incompatibility within the elements of a
provided class list. (Here compatible means "non-overlapping".)

- Necessary changes in the methods of Class.java and ClassBundle.java
have been made so as to successfully implement the methods above.