Button for importing DB from a selected .csv file is added & implemented
- "PrepareNormalizedTablesFromCSV" method can now use a file chooser
to select a .csv file which is then used to fill the .db accordingly.

- PrepareNormalizedTablesFromCSV is run once at the initialization with
a default .csv file corresponding to "Fall 2019", and then user is given
the opportunity to select an external .csv file which is to be used in
the re-generation of the current .db
(i.e. it overwrites the current .db by using the provided .csv file)

- User is advised to use a UTF-8 compatible .csv file before picking
the desired .csv file

- ShowMessage method is divided into 2 different methods
(e.g. ShowMessageNow and ShowMessageLater)

- Changes have been made in the following files:

         - Main.java (modified)