Major GUI Bugs Have Been Fixed
- Now, the pane for weekly schedule GUI doesn't get closed each time
user clicks next or prev. buttons. Instead, it now repaints to the same
pane whenever user wants to view next/prev. schedule.
(i.e. first clears the components of the frame other than buttons, and then
it draws the next/prev. schedule. At last, it repaints the frame so as to
make changes visible.)

- Another GUI bug, which used to occur upon applying filters using "Apply
Filters" button, is fixed.
(e.g. view weekly time table pane was sent back due to the prompt message
of filtering info. This is now prevented by setting the parent of this prompt
message to "null".)

 - Changes are made in the following classes

          -Main
          -WeeklyScheduleGUI