Weekly Timetable GUI now shows 19.00
- Improvements on ClassPanel & BackgroundPanel are done such that
they are now more dynamic (i.e. they are receiving parameters to their
constructors and doing calculations in accordance with these as opposed
to using static integer values.)

- Weekly Timetable GUI now shows 19.00 as its final time stamp.
(e.g. it handles classes that start at 17.30 and end at 18.45)

- The buttons on the Weekly Timetable GUI are now located at a specific Y
coordinate that is calculated through a dynamic Y_offset parameter
(i.e. a certain value of Y_offset from the bottom of the frame.)

 - Changes are made in the following classes

          -WeeklyScheduleGUI
          -BackgroundPanel
          -ClassPanel