UPDATE for WeeklyScheduleTimeTable's Title Text
- WeeklyScheduleTimeTable now also shows "Plan #<PlanNumber>" in its title

- Bug fix for WeeklyScheduleTimeTable where it used to open the latest frame
from the last run even though having applied a filter which yields 0 schedules.
(i.e. although it shouldn't, it used to open a frame belonging to the previous run)
(e.g. thread.join() is used in the ApplyFilters button which makes the main thread
wait until thread for applying filters is done.)

- cleanText() method is added to WeeklyScheduleTimeTable class which cleans
the selected courses text from unnecessary characters such as '{', '}' and '\"'
(i.e. it is used in the text generation for the title of WeeklyScheduleTimeTable.)

- SaveAsImage function in Parser class now doesn't print any unnecessary information
regarding schedule and its respective schedule id.
(i.e. it only prints "Plan #<no>" for each .jpeg file generated.)

- Changes are made in the following classes:

        - Main.java (modified)
        - WeeklyScheduleTimeTable.java (modified)
        - Parser.java (modified)